refactor(start): break the config file into smaller pieces
